% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-group-functions.R
\name{add_age_cs_state_col}
\alias{add_age_cs_state_col}
\title{add_age_cs_state_col
Add in the age_cs_state column to a data frame}
\usage{
add_age_cs_state_col(my_df)
}
\arguments{
\item{my_df}{data frame}
}
\description{
add_age_cs_state_col
Add in the age_cs_state column to a data frame
}
