% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-group-functions.R
\name{add_age_group_column}
\alias{add_age_group_column}
\title{Add age group column to a data frame
using ONS standards}
\usage{
add_age_group_column(my_tbl, age_column_name = "age")
}
\arguments{
\item{my_tbl}{a tibble with the column age}

\item{age_column_name}{a character vector of the name of the age column}
}
\description{
Add age group column to a data frame
using ONS standards
}
