% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/params-numbers-births-migrations-deaths-function.R
\name{calc_death_prop_age}
\alias{calc_death_prop_age}
\title{Calculate the Percentage of Deaths which are from the population aged between two ages for a given period}
\usage{
calc_death_prop_age(
  age_lower_bound = 0,
  age_upper_bound = Inf,
  sql_con = NA,
  start_date = "2020-01-01",
  end_date = "2020-12-31"
)
}
\arguments{
\item{age_lower_bound}{numeric the minimum age for death, 0 if left blank}

\item{age_upper_bound}{numeric the maximum age for death, inf if left blank}

\item{sql_con}{connection to SQL database}

\item{start_date}{character format YYYY-MM-DD}

\item{end_date}{character format YYYY-MM-DD}
}
\description{
Calculate the Percentage of Deaths which are from the population aged between two ages for a given period
}
