% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN-run-dpm-function-age-based.R
\name{calculate_weighted_birth_im_em_death_probs}
\alias{calculate_weighted_birth_im_em_death_probs}
\title{function to calculate an adjusted probability matrix for one of immigration/emigration/births
given the current population distribution}
\usage{
calculate_weighted_birth_im_em_death_probs(
  event_probs,
  current_pop,
  number_of_events,
  chosen_event,
  simple_output = T
)
}
\arguments{
\item{event_probs}{the original event probabilities expected per person}

\item{current_pop}{the current population distribution}

\item{number_of_events}{the total number of events we need to decide}

\item{chosen_event}{one of 'births','immigrations','emigrations'}

\item{simple_output}{boolean, if TRUE returns minimal cols, with column value as key metric}
}
\description{
function to calculate an adjusted probability matrix for one of immigration/emigration/births
given the current population distribution
}
