% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner-transition-functions.R
\name{check_inner_trans}
\alias{check_inner_trans}
\title{check the inner transition matrix is valid, and if it is only a matrix convert it
into a list of size total_time where every element is the same matrix}
\usage{
check_inner_trans(inner_trans_matrix_list, total_time, warnings = TRUE)
}
\arguments{
\item{inner_trans_matrix_list}{the inner transition matrix list, or perhaps matrix}

\item{total_time}{the time window}

\item{warnings}{T/F - whether to output warning messages}
}
\description{
check the inner transition matrix is valid, and if it is only a matrix convert it
into a list of size total_time where every element is the same matrix
}
