% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslag-sem.R
\name{clfe}
\alias{clfe}
\title{Cross-lagged panel models with fixed effects}
\usage{
clfe(formula, data, id = NULL, wave = NULL, err_inv = FALSE,
  const_inv = FALSE, alpha_free = FALSE, last.wave = FALSE,
  print_only = FALSE, ...)
}
\arguments{
\item{formula}{Model formula. See details for instructions on specifying
parameters properly.}

\item{data}{Data frame in "long" format. Can be a "panel_data" object.}

\item{id}{Name of the data column that identifies which individual the
observation is. Not needed if \code{data} is a "panel_data" object.}

\item{wave}{Name of the data column that identifies which wave the
observation is from. Not needed if \code{data} is a "panel_data" object.}

\item{err_inv}{Constrain the error variance to be equal across waves.
Default is FALSE.}

\item{const_inv}{Constrain the dependent variable's variance to be equal
across waves. This removes cross-sectional dependence. Default is FALSE.}

\item{alpha_free}{Estimate each wave of the dependent variable's loading on
the alpha latent variable. Default is FALSE, meaning each wave has a loading
of 1.}

\item{last.wave}{Use the final wave of lagged predictors? Default is FALSE,
but enables better specification of 3-wave lagged models.}

\item{print_only}{Instead of estimating the model, print the \pkg{lavaan}
model string to the console instead.}

\item{...}{Extra parameters to pass \code{\link[lavaan]{sem}}. Examples
could be \code{missing = "fiml"} for missing data or
\code{estimator = "MLM"} for robust estimation.}
}
\value{
An object of class "clfe," which has its own
 \code{\link[clfe]{summary}} method.

 The clfe object also returns: \itemize{

 \item \code{fit}, the \code{\link[lavaan]{lavaan}} object fit by this
  function. This could be useful if you want to use lavaan's functions to
  further learn from the model
 \item \code{mod_string}, the character object used to specify the model
  to lavaan. This is helpful if you want to fit the model yourself or
  wish to check that the specification is correct.
 \item \code{wide_data}, the widened data frame necessary to fit the SEM.

 }
}
\description{
Estimate dynamic panel models with fixed effects via
maximum likelihood estimation.
}
\examples{
# Load example data
data("WageData", package = "clfe")
# Convert data to panel_data format for ease of use
wages <- panel_data(WageData, id = id, wave = t)

# Replicates Allison, Williams, & Moral-Benito (2017) analysis
fit <- clfe(wks ~ pre(lag(union)) + lag(lwage) | ed, data = wages,
            err_inv = TRUE, information = "observed")
# Note: information = "observed" only needed to match Stata/SAS standard errors
summary(fit)


}
\author{
Jacob A. Long, in consultation with Richard A. Williams. All errors
 are Jacob's.
}
