% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/params-numbers-births-migrations-deaths-function.R
\name{convert_yearly_to_monthly}
\alias{convert_yearly_to_monthly}
\title{subfunction for converting output of get_numbers_births_migrations_deaths
into one row being a month, from one row being a year. Values are split equally
into 12}
\usage{
convert_yearly_to_monthly(figures_tbl)
}
\description{
subfunction for converting output of get_numbers_births_migrations_deaths
into one row being a month, from one row being a year. Values are split equally
into 12
}
