% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{core_seg_cols_greenred}
\alias{core_seg_cols_greenred}
\title{core_seg_cols_greenred}
\format{
## `core_seg_cols_greenred`
A named vector with values
\describe{
  \item{name}{Core Segment eg 'CS1'}
  \item{value}{Hex code eg '#77A033'}
}
}
\usage{
core_seg_cols_greenred
}
\description{
core_seg_cols_greenred
}
\keyword{datasets}
