% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compound-growth-functions.R
\name{create_growth_tbl}
\alias{create_growth_tbl}
\title{A subfunction of plot_dpm_with_growth.}
\usage{
create_growth_tbl(num_years, growth_factor, start_val)
}
\arguments{
\item{num_years}{number of years in data}

\item{growth_factor}{1 is no growth, 1.1 is yearly 10\% compound growth}

\item{start_val}{the starting value to grow from}
}
\description{
A subfunction of plot_dpm_with_growth.
}
