% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sankey-functions.R
\name{create_sankey}
\alias{create_sankey}
\title{create Sankey diagram from a DPM run}
\usage{
create_sankey(
  population_at_each_year,
  inner_trans_matrix_list,
  save = FALSE,
  save_filename = NA
)
}
\arguments{
\item{population_at_each_year}{output from dpm::run_dpm}

\item{inner_trans_matrix_list}{inner transition matrix}

\item{save}{logical whether to save output as png}

\item{save_filename}{default NA, what filepath should output be. Only works if save=T}
}
\description{
create Sankey diagram from a DPM run
}
