% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from-population-to-other-metrics.R
\name{create_type_forecasts_from_population}
\alias{create_type_forecasts_from_population}
\title{Create data forecasting given type_name with the DPM output}
\usage{
create_type_forecasts_from_population(
  dpm_output,
  core_segment_by_type,
  type_name
)
}
\arguments{
\item{dpm_output}{dataframe output from run_dpm function}

\item{type_name}{char character name for what you want the value of the type
column to be called in the output ie what is the measure in core_segments_by_type}

\item{core_segments_by_type}{dataframe with columns state_name and yearly_amount}
}
\description{
Given the dpm_output from run_dpm, and the breakdown
of core_segment_by_type showing the amount each type_name used in a year
by the average person in that Core Segment, create the forecast.
}
\details{
Output can be plotted with plot_dpm_with_growth function
}
