% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_dpm.R
\docType{class}
\name{dpm-class}
\alias{dpm-class}
\title{Dynamic Panel Model (\code{dpm}) class}
\description{
Models fit using \code{\link[=dpm]{dpm()}} return values of this class, which
inherits from \code{\link[lavaan]{lavaan-class}}.
}
\section{Slots}{

\describe{
\item{\code{call_info}}{A list of metadata about the arguments used.}

\item{\code{call}}{The actual function call.}

\item{\code{mod_string}}{The model formula passed to \code{lavaan}.}

\item{\code{wide_data}}{The data provided to the \code{data} argument in the function
call.}
}}

