% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPM-package.R
\docType{package}
\name{DPM-package}
\alias{DPM}
\alias{DPM-package}
\title{DPM: Dirichlet Process Mixture for Polygenic Risk Decomposition}
\description{
DPM integrates gene-level GWAS response statistics with gene-by-label
single-cell expression profiles to infer latent gene modules and
module-specific cellular coefficients.
}
\author{
\strong{Maintainer}: Jiacheng Lou \email{jiachenglou82@gmail.com}

}
\keyword{internal}
