% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-format-functions.R
\name{from_list_to_long_tbl}
\alias{from_list_to_long_tbl}
\title{same as from_matrix_to_long_tbl but for list input}
\usage{
from_list_to_long_tbl(inner_trans_matrix_list)
}
\arguments{
\item{inner_trans_matrix_list}{list of inner transition matrices}
}
\description{
same as from_matrix_to_long_tbl but for list input
}
