% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-format-functions.R
\name{from_matrix_to_long_tbl}
\alias{from_matrix_to_long_tbl}
\title{takes inner_trans_matrix and makes it a long tibble object}
\usage{
from_matrix_to_long_tbl(inner_trans_matrix)
}
\arguments{
\item{inner_trans_matrix}{inner transition matrix}
}
\description{
takes inner_trans_matrix and makes it a long tibble object
}
