% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN-run-dpm-function-age-based.R
\name{get_age_based_dpm_inputs_from_folder}
\alias{get_age_based_dpm_inputs_from_folder}
\title{function for reading in all .RDS inputs of age-based DPM into}
\usage{
get_age_based_dpm_inputs_from_folder(folder)
}
\arguments{
\item{folder}{file path for folder inputs from DPMmicrosim age-based model}
}
\description{
function for reading in all .RDS inputs of age-based DPM into
}
