% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/params-proportion-births-migrations-deaths-function.R
\name{get_births_migrations_deaths_proportions}
\alias{get_births_migrations_deaths_proportions}
\title{Given a start_month, compare the population then VS compare_against months prior,
to output the proportion of births, migrations, and deaths that come from
each Core Segment}
\usage{
get_births_migrations_deaths_proportions(
  start_month,
  source_or_preload = "source",
  method = 1,
  sql_con = NA,
  min_age = 17,
  compare_against = 12,
  output_proportions_or_numbers = "proportions"
)
}
\arguments{
\item{start_month}{the start month as a character of form YYYY-MM}

\item{source_or_preload}{either 'source' or 'preloaded' depending on whether to
calculate source_or_preload the source SQL tables or use a pre-created file}

\item{method}{either a number (1 default) or string specifying the method. Valid
methods are
1  or "Matching at patient-level migrations in/out"
2  or "Taking average movements from data"}

\item{sql_con}{connection to SQL database}

\item{min_age}{the minimum age to be included in the model}

\item{output_proportions_or_numbers}{should the output tibble be raw numbers, or
proportions}
}
\description{
Given a start_month, compare the population then VS compare_against months prior,
to output the proportion of births, migrations, and deaths that come from
each Core Segment
}
