% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/params-proportion-births-migrations-deaths-function.R
\name{get_bmd_vals_matching_method}
\alias{get_bmd_vals_matching_method}
\title{subfunction of get_births_migrations_deaths_proportions when method is
1 or "Matching at patient-level migrations in/out}
\usage{
get_bmd_vals_matching_method(
  sql_con,
  start_month_date_char,
  compare_against_month_date_char,
  min_age = 17
)
}
\description{
subfunction of get_births_migrations_deaths_proportions when method is
1 or "Matching at patient-level migrations in/out
}
