% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-initial-population-function.R
\name{get_initial_population}
\alias{get_initial_population}
\title{Given a month, create the initial_population tibble input, by Core Segment}
\usage{
get_initial_population(
  start_month,
  source_or_preload = "source",
  method = 1,
  sql_con = NA,
  min_age = 17,
  age_groups = F
)
}
\arguments{
\item{start_month}{the start month as a character of form YYYY-MM}

\item{source_or_preload}{either 'source' or 'preloaded' depending on whether to
calculate source_or_preload the source SQL tables or use a pre-created file}

\item{method}{either a number (1 default) or string specifying the method. Valid
methods are
1  or "CS props: Cleaned CMS values. Total pop: GP Estimates scaled down 90% to match ONS"
2  or "CS props: Cleaned CMS values. Total pop: Cleaned CMS Values"
3  or "CS props: Raw CMS values. Total pop: GP Estimates scaled down 90% to match ONS"}

\item{sql_con}{connection to SQL database}

\item{min_age}{the minimum age to be included in the model}

\item{age_groups}{boolean, whether to separate out by age groups}
}
\description{
Given a month, create the initial_population tibble input, by Core Segment
}
