% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/params-numbers-births-migrations-deaths-function.R
\name{get_numbers_births}
\alias{get_numbers_births}
\title{get the numbers of births from ONS projections}
\usage{
get_numbers_births(input_folder_loc)
}
\description{
get the numbers of births from ONS projections
}
