% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/params-numbers-births-migrations-deaths-function.R
\name{get_numbers_births_migrations_deaths}
\alias{get_numbers_births_migrations_deaths}
\title{only way we have to get predictions is from ONS figures}
\usage{
get_numbers_births_migrations_deaths()
}
\description{
only way we have to get predictions is from ONS figures
}
