% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/params-numbers-births-migrations-deaths-function.R
\name{get_ons_forecast_variant_options}
\alias{get_ons_forecast_variant_options}
\title{Show what options exist for ONS forecasts}
\usage{
get_ons_forecast_variant_options(
  url = "I don't know the url",
  message_with_options = TRUE
)
}
\arguments{
\item{url}{the URL, can be left blank}

\item{message_with_options}{logical whether to print the options to the console}
}
\description{
Show what options exist for ONS forecasts
}
