% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-initial-population-function.R
\name{get_pop_from_gp_data}
\alias{get_pop_from_gp_data}
\title{subfunction of get_initial_population when method is
1 or "CS props: Cleaned CMS CS. Total pop: GP Estimates scaled down 90% to match ONS"}
\usage{
get_pop_from_gp_data(start_month_date_char, sql_con, min_age = 17)
}
\description{
subfunction of get_initial_population when method is
1 or "CS props: Cleaned CMS CS. Total pop: GP Estimates scaled down 90% to match ONS"
}
