% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-functions.R
\name{get_sql_con}
\alias{get_sql_con}
\title{connection to the SQL server}
\usage{
get_sql_con(server = NA)
}
\arguments{
\item{server}{string name of the Server to connect to. If not entered, default
is to look for 'Server' environment variable}
}
\description{
connection to the SQL server
}
