% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpm.R
\name{get_syntax}
\alias{get_syntax}
\title{Retrieve lavaan model syntax from fitted dpm model}
\usage{
get_syntax(model, print = TRUE)
}
\arguments{
\item{model}{A \code{dpm} object.}

\item{print}{Print the syntax to the console so it is formatted properly?
Default is TRUE.}
}
\description{
This helper function provides a simple way to retrieve the
lavaan model syntax from a fitted \code{\link[dpm:dpm]{dpm::dpm()}} object.
}
