% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-get-inner-transition-matrix-functions.R
\name{get_transition_numbers}
\alias{get_transition_numbers}
\title{Calculate all the transitions between two months}
\usage{
get_transition_numbers(
  sql_con,
  orig_month_start_date,
  method = 1,
  compare_against = 12,
  min_age = 17
)
}
\arguments{
\item{sql_con}{connection to SQL database}

\item{orig_month_start_date}{month interested in date format, day of month is not considered}

\item{method}{options for which method to use}

\item{compare_against}{the number of months back to look}

\item{min_age}{the minimum age to consider. Default 17}
}
\description{
Calculate all the transitions between two months
}
