% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpm.R
\name{get_wide_data}
\alias{get_wide_data}
\title{Retrieve wide-format data from fitted dpm model}
\usage{
get_wide_data(model)
}
\arguments{
\item{model}{A \code{dpm} object.}
}
\description{
This helper function provides a simple way to retrieve the
widened data from a fitted \code{\link[dpm:dpm]{dpm::dpm()}} object.
}
