% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-format-functions.R
\name{make_links_df}
\alias{make_links_df}
\title{create a links_df object for use in plotting a sankey}
\usage{
make_links_df(population_at_each_year, inner_trans_matrix_list)
}
\arguments{
\item{population_at_each_year}{output of dpm::run_dpm()}

\item{inner_trans_matrix_list}{inner transition matrix}
}
\description{
create a links_df object for use in plotting a sankey
}
