% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/META-full-dpm.R
\name{meta_run_dpm}
\alias{meta_run_dpm}
\title{meta_run_dpm runs the entire DPM, using subfunctions from within this package
in such a way that the inputs to this function fully determine the outputs}
\usage{
meta_run_dpm(
  start_month,
  continuous_dpm = T,
  source_or_preload = "source",
  get_initial_population_method =
    "CS props: Raw CMS values. Total pop: GP Estimates scaled down 90\% to match ONS",
  get_transition_numbers_method = "Using only patients with full records of data",
  get_numbers_births_migrations_deaths_method = "use ONS closest year",
  get_births_migrations_deaths_proportions_method =
    "Matching at patient-level migrations in/out",
  combine_immigration_emigration = F,
  min_age = 17,
  age_groups = T,
  ons_forecast_variant = "normal",
  my_seed = 1,
  save_folder = NA
)
}
\description{
meta_run_dpm runs the entire DPM, using subfunctions from within this package
in such a way that the inputs to this function fully determine the outputs
}
