% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdata.R
\name{panel_data}
\alias{panel_data}
\title{Make data frames into panel data frames.}
\usage{
panel_data(data, id, wave)
}
\arguments{
\item{data}{A data frame or similar in "long" format, with each row
representing a single individual at a single time.}

\item{id}{The name of the column with individual identifiers. Input can be
a string or unquoted name.}

\item{wave}{The name of the column with time/wave identifiers. Input can be
a string or unquoted name.}
}
\value{
A data frame with class "panel_data".
}
\description{
Panel data frames are unchanged with the exception of added
 attributes for better compatibility with longitudinal analysis functions.
}
\examples{
data("WageData", package = "clfe")
wages <- panel_data(WageData, id = id, wave = t)

}
