% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-based.R
\name{plot_age_based_blobs}
\alias{plot_age_based_blobs}
\title{given an input_tbl with columns containing age_cs_state_col - plot them as blobs
using the value_col column for the blob sizing}
\usage{
plot_age_based_blobs(
  input_tbl,
  age_cs_state_col = "age_cs_state",
  value_col = "value",
  blob_labels = T,
  label_type = "all"
)
}
\description{
given an input_tbl with columns containing age_cs_state_col - plot them as blobs
using the value_col column for the blob sizing
}
