% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-transition-matrix.R
\name{plot_inner_transition_matrix}
\alias{plot_inner_transition_matrix}
\title{create Inner Transition Matrix as a plot}
\usage{
plot_inner_transition_matrix(
  inner_trans_matrix_list,
  transition_to_plot = 1,
  save = FALSE,
  save_filename = NA
)
}
\arguments{
\item{inner_trans_matrix_list}{inner transition matrix}

\item{transition_to_plot}{integer, which transition year to plot for}

\item{save}{logical whether to save output as png}

\item{save_filename}{default NA, what filepath should output be. Only works if save=T}
}
\description{
create Inner Transition Matrix as a plot
}
