% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/params-numbers-births-migrations-deaths-function.R
\name{read_ons_projections_online}
\alias{read_ons_projections_online}
\title{Given a forecast variant from ONS, read the appropriate Excel file and clean
it, into a tidy format for use elsewhere}
\usage{
read_ons_projections_online(forecast_variant, table, sleep = 1)
}
\arguments{
\item{forecast_variant}{character the variant of the forecast you want to use}

\item{table}{character the table you want to use, one of:
- 'births_deaths_migration' the table of births, deaths and migration.
- 'population_by_age' the table of population by 5 year age bands. Useful for
births' of 17+}

\item{sleep}{the number of seconds to pause for, to avoid errors for requestion too much from the web}
}
\description{
Given a forecast variant from ONS, read the appropriate Excel file and clean
it, into a tidy format for use elsewhere
}
