% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{panel_data}
\alias{long_panel}
\alias{widen_panel}
\alias{get_wave}
\alias{get_id}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{panelr}{\code{\link[panelr]{panel_data}}, \code{\link[panelr]{long_panel}}, \code{\link[panelr]{widen_panel}}, \code{\link[panelr]{get_wave}}, \code{\link[panelr]{get_id}}}
}}

