% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN-run-dpm-function-continuous-time.R
\name{run_dpm_ct}
\alias{run_dpm_ct}
\title{Main DPM function under continuous time, which is what the _ct stands for in
the function name.}
\usage{
run_dpm_ct(
  initial_population,
  inner_trans_matrix,
  monthly_entrants_exits,
  start_time = 0,
  seed = 1
)
}
\description{
Main DPM function under continuous time, which is what the _ct stands for in
the function name.
}
