% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner-transition-functions.R
\name{scalar_from_to}
\alias{scalar_from_to}
\title{takes an Inner Transition Matrix and which inner transition to rescale by scalar_change, for example
a 10% reduction would be scalar_change = 0.9}
\usage{
scalar_from_to(
  inner_trans_matrix,
  from_cs,
  to_cs,
  scalar_change,
  method = "take from no change"
)
}
\arguments{
\item{inner_trans_matrix}{matrix input}

\item{from_cs}{integer. The Core Segment to change from. Equivalent to *row* of inner_trans_matrix.}

\item{to_cs}{integer. The Core Segment to change to. Equivalent to *column* of inner_trans_matrix.}

\item{scalar_change}{numeric, the amount to scale the transition from from_cs to to_cs}

\item{method}{either "take from no change" or "take proportionally from other changes" - how to get the
proportions to add back to 1}
}
\description{
takes an Inner Transition Matrix and which inner transition to rescale by scalar_change, for example
a 10% reduction would be scalar_change = 0.9
}
