% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner-transition-functions.R
\name{valid_inner_trans_matrix}
\alias{valid_inner_trans_matrix}
\title{checks whether the inner transition matrix is valid - outputs errors or
warnings if not}
\usage{
valid_inner_trans_matrix(inner_trans_matrix, warnings = T)
}
\arguments{
\item{inner_trans_matrix}{an inner transition matrix}
}
\description{
checks whether the inner transition matrix is valid - outputs errors or
warnings if not
}
