% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-get-inner-transition-matrix-functions.R
\name{create_many_monthly_transitions}
\alias{create_many_monthly_transitions}
\title{applies create_monthly_transition_rate over multiple months}
\usage{
create_many_monthly_transitions(
  sql_con,
  months_to_go_through,
  compare_against = 12,
  min_age = 17,
  method = 1
)
}
\arguments{
\item{sql_con}{connection to Microsoft SQL server object}

\item{months_to_go_through}{vector of months to go through}

\item{compare_against}{the number of months back to look}

\item{min_age}{minimum age of patient}

\item{method}{an integer, see get_transition_numbers}
}
\description{
applies create_monthly_transition_rate over multiple months
}
