% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpm.R
\docType{methods}
\name{update}
\alias{update}
\alias{update.dpm}
\alias{update,dpm-method}
\alias{show,dpm-method}
\alias{coef}
\alias{coef.dpm}
\alias{coef,dpm-method}
\alias{formula,dpm-method}
\title{Various methods for \code{dpm} objects}
\usage{
update(object, ...)

\method{update}{dpm}(object, formula., ..., evaluate = TRUE)

\S4method{update}{dpm}(object, formula., ..., evaluate = TRUE)

\S4method{show}{dpm}(object)

coef(object, ...)

\method{coef}{dpm}(object)

\S4method{coef}{dpm}(object)

\S4method{formula}{dpm}(x)
}
\arguments{
\item{object}{A \code{dpm} object}

\item{...}{Other arguments to update.}

\item{formula.}{An updated formula (optional)}

\item{evaluate}{If updating, should the updated model be updated or just
return the call? Default is TRUE, re-run the model.}

\item{x}{A \code{dpm} object}
}
\description{
R likes it when these things have documentation.
}
