% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/params-proportion-births-migrations-deaths-function.R
\name{get_bmd_vals_matching_method}
\alias{get_bmd_vals_matching_method}
\title{subfunction of get_births_migrations_deaths_proportions when method is
1 or "Matching at patient-level migrations in/out}
\usage{
get_bmd_vals_matching_method(
  sql_con,
  start_month_date_char,
  compare_against_month_date_char,
  combine_immigration_emmigration = TRUE,
  min_age = 17
)
}
\arguments{
\item{sql_con}{connection to SQL database}

\item{start_month_date_char}{character string format YYYY-MM-DD}

\item{compare_against_month_date_char}{character string format YYYY-MM-DD}

\item{combine_immigration_emmigration}{boolean whether to combine into migration or not}

\item{min_age}{default 17}
}
\description{
subfunction of get_births_migrations_deaths_proportions when method is
1 or "Matching at patient-level migrations in/out
}
