% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-get-inner-transition-matrix-functions.R
\name{get_inner_trans_rate_from_transitions_tbl}
\alias{get_inner_trans_rate_from_transitions_tbl}
\title{takes output from get_transition_numbers and creates the inner transitions matrix
out_type can either be 'matrix' or 'tbl"}
\usage{
get_inner_trans_rate_from_transitions_tbl(transitions, out_type = "matrix")
}
\arguments{
\item{transitions}{tibble output from get_transition_numbers}

\item{out_type}{either 'matrix' or 'tbl' depending on your flAvah}
}
\description{
takes output from get_transition_numbers and creates the inner transitions matrix
out_type can either be 'matrix' or 'tbl"
}
