% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clfe.R
\name{lav_summary}
\alias{lav_summary}
\alias{lav_fitmeasures}
\title{lavaan interface to clfe objects}
\usage{
lav_summary(x, ...)

lav_fitmeasures(x, ...)
}
\arguments{
\item{x}{The \code{\link{clfe}} object}

\item{...}{Other arguments to the the lavaan function.}
}
\description{
These are essentially shortcuts to commonly-desired lavaan
 functions to be used with clfe objects, which contain a lavaan model.
}
\details{
Here are the currently supported functions with a brief explantion:

  \itemize{

    \item \code{lav_summary}: lavaan's summary function.
    \item \code{lav_fitmeasures}: lavaan's \code{summary}

  }
}
\examples{

# Load example data
data("WageData")
# Convert data to panel_data format for ease of use
wages <- panel_data(WageData, id = id, wave = t)

fit <- clfe(wks ~ pre(lag(union)) + lag(lwage) | ed, data = wages)
lav_summary(fit)
lav_fitmeasures(fit)

}
\references{
Allison, P. D., Williams, R., & Moral-Benito, E. (2017). Maximum likelihood
for cross-lagged panel models with fixed effects. *Socius*, *3*, 1–17.
http://journals.sagepub.com/doi/10.1177/2378023117710578
}
