% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compound-growth-functions.R
\name{plot_dpm_with_growth}
\alias{plot_dpm_with_growth}
\title{plot the DPM outputs with growth factors
import @dplyr
import @tidyr
import @ggplot2}
\usage{
plot_dpm_with_growth(dpm_output, growth_factors = c(1.01, 1.02, 1.03))
}
\description{
plot the DPM outputs with growth factors
import @dplyr
import @tidyr
import @ggplot2
}
