% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN-run-dpm-function.R
\name{run_dpm}
\alias{run_dpm}
\title{run_dpm main function for running the dpm model}
\usage{
run_dpm(
  initial_population,
  inner_trans_matrix_list,
  total_time,
  births_migrations_deaths_figures,
  birth_migration_deaths_proportions
)
}
\arguments{
\item{initial_population}{initial population tibble}

\item{inner_trans_matrix_list}{list of inner transition matrices}

\item{total_time}{integer}

\item{births_migrations_deaths_figures}{tibble}

\item{birth_migration_deaths_proportions}{tibble}
}
\description{
run_dpm main function for running the dpm model
}
