% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN-run-dpm-function-age-based.R
\name{run_dpm_age_based}
\alias{run_dpm_age_based}
\title{run_dpm main function for running the dpm model}
\usage{
run_dpm_age_based(
  folder,
  print_intermediate_results = T,
  return_yearly_deaths = F,
  return_yearly_changes = F
)
}
\arguments{
\item{folder}{folder to find all information from}

\item{print_intermediate_results}{boolean if TRUE then outputs as it goes
returned with named outputs}

\item{return_yearly_deaths}{boolean if TRUE then returns a table of deaths}

\item{return_yearly_changes}{boolean if TRUE then returns a table of aggregated changes to the population}
}
\description{
takes folder input as used in DPMmicrosim terminology for outputs, so is
compatible with that ecosystem
}
