% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN-run-dpm-function-age-based.R
\name{run_dpm_age_based}
\alias{run_dpm_age_based}
\title{run_dpm main function for running the dpm model}
\usage{
run_dpm_age_based(folder, print_intermediate_results = T)
}
\arguments{
\item{folder}{folder to find all information from}

\item{print_intermediate_results}{boolean if TRUE then outputs as it goes
returned with named outputs}
}
\description{
takes folder input as used in DPMmicrosim terminology for outputs, so is
compatible with that ecosystem
}
