% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-inputs-outputs.R
\name{save_dpm_inputs_outputs}
\alias{save_dpm_inputs_outputs}
\title{save all the inputs and outputs - pulls from environment variables}
\usage{
save_dpm_inputs_outputs(save_loc = here::here())
}
\description{
save all the inputs and outputs - pulls from environment variables
}
